refactor: add .sequelizerc and configure models inside src
