+++ backend/.sequelizerc
@@ -1,0 +1,6 @@
+const path = require("path");
+
+module.exports = {
+    'config': path.resolve('config', 'config.js'),
+    'models-path': path.resolve('src', 'models'),
+};
