+++ backend/src/repository/userRepository.ts
@@ -1,0 +1,3 @@
+import db from "../models";
+
+export class UserRepository {}
